move `OrderedCache` to `common`
