+++ src/stream/src/common/cache/mod.rs
@@ -12,10 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use builder::*;
-pub use column_mapping::*;
+mod ordered_cache;
 
-mod builder;
-pub mod cache;
-mod column_mapping;
-pub mod table;
+pub use ordered_cache::*;
